+++ news_aggregator/server/services/user_article_service.py
@@ -1,6 +1,5 @@
 import logging
 from server.db.database import get_db
-from datetime import datetime
 
 
 class UserArticleService:
